+++ linza_mcp/compat.py
@@ -97,7 +97,7 @@ from .utils import (
 )
 
 
-__version__ = "0.1.3"
+__version__ = "0.1.4"
 
 
 class LinzaCore: